requirements.txt
absl-py==1.0.0
arrow==1.2.1
astunparse==1.6.3
async-generator==1.10
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cookiecutter==1.7.3
Flask==2.0.2
Flask-BasicAuth==0.2.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
idna==3.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jinja2-time==0.2.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.4
Pillow==8.4.0
poyo==0.5.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjson==1.3.0
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.2
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.0.0
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
xgboost==1.5.0
gunicorn==20.1.0
